requirement.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.8
decorator==4.4.2
edge-tts==6.1.12
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl#sha256=61e6e5530941f5880166855f09f60d7e6ba79ec1e8e45f96244bdb1eb169eb1d
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.23.4
hyperframe==5.2.0
idna==2.10
imageio==2.34.2
imageio-ffmpeg==0.5.1
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
moses==0.10.0
moviepy==1.0.3
mpmath==1.3.0
mpv==1.0.7
multidict==6.0.5
murmurhash==1.0.10
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
preshed==3.0.9
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.7.1
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.3
scipy==1.14.0
selenium==4.23.0
sentencepiece==0.2.0
setuptools==70.3.0
shellingham==1.5.4
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.5
spacy-alignments==0.9.1
spacy-curated-transformers==0.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
spacy-pkuseg==0.0.33
spacy-transformers==1.3.5
srsly==2.4.8
sympy==1.13.0
thinc==8.2.5
thinc_apple_ops==0.1.5
tokenizers==0.15.2
torch==2.3.1
tqdm==4.66.4
transformers==4.36.2
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl#sha256=f51075665749e07406d629d1055ce5a68635fae6ab3c34257ee798c62b4fc431
zh-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_trf-3.7.2/zh_core_web_trf-3.7.2-py3-none-any.whl#sha256=16b8d4bf23d20a04cfcbe676ae1be2be4437b40cf8101c9f3e7f6db4674ec91d
